requirements.txt
streamlit>=1.31.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
ffmpeg>=1.4 
python-docx>=1.0.1
pdfkit>=1.0.0
pytube==15.0.0
moviepy>=1.0.3
nest-asyncio==1.5.8
torch>=2.0.0
requests==2.31.0
transformers==4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
yt-dlp>=2023.12.30
numpy>=1.20.0